♻ Update @percy/client exports
✅ Rename client test helper
