+++ packages/client/package.json
@@ -10,13 +10,18 @@
   "publishConfig": {
     "access": "public"
   },
-  "main": "dist/index.js",
-  "files": [
-    "dist",
-    "test/helpers.js"
-  ],
   "engines": {
     "node": ">=12"
+  },
+  "files": [
+    "./dist",
+    "./test/helpers.js"
+  ],
+  "main": "./dist/index.js",
+  "exports": {
+    ".": "./dist/index.js",
+    "./utils": "./dist/utils.js",
+    "./test/helpers": "./test/helpers.js"
   },
   "scripts": {
     "build": "node ../../scripts/build",
@@ -27,8 +32,5 @@
   "dependencies": {
     "@percy/env": "1.0.0-beta.76",
     "@percy/logger": "1.0.0-beta.76"
-  },
-  "devDependencies": {
-    "mock-require": "^3.0.3"
   }
 }
